properformat logic
I know this commit is more than 5 lines. But this is just simply copied and pasted from past projects. We are just checking if the char is in PROPER RANGE. And looping through each character. If it is not we simply go to INVALID.

NOTE: We change t5 to 21 because we can accept only 20 CHARACTERS this time!