done w/ checking tabs and spaces
now that we are done, we will continue with our program. we quickly check if we haven't counted any characters.then just jump to invalid. Then:
We have to initialize important things for our JAL recursive function.
base case checker -- init to zero to determine if we reached our base case
Base case checker -- copy base case checker to s6 so that we can decrease it everytime we call our function
multiply controller -- if we have "ABC" the max we need to multiply is 2 times.
copy multiply controller -- used to multiply our d cMC number of times; we will use this; increasing the amount we multiply