Print Result Protocol -- IMPORTANT PATTERN
Once we are done with our function we need to print the result. This is something that is NEW different than other programs.

Now we have our finalAnswer output is holding the answer but it is reversed. This is something that I take advantage of throughout my program. Since we limit ourselves to 20 characters, we know the # of digits in the output will never be bigger than 50. So throughout our calculation we will take advanatge of this. 

To print this. We SAVE the ADDRESS OF THE LAST NON-ZERO number.  Once we have that we can just print the values until the base address == that address. 

For example: Input: "A"; Output: 000000000000000000000100. We find save the address of 1 because the rest is zeros. And print: 100